pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "Private REST API"
authors = ["Antoine Lancial <antoinelncl@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
sqlmodel = "^0.0.18"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
slowapi = "^0.1.9"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
itsdangerous = "^2.2.0"
debugpy = "^1.8.1"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
flake8-type-checking = "^2.9.0"
flake8-fastapi = "^0.7.0"
flake8-bugbear = "^24.4.26"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
faker = "^25.2.0"
coverage = "^7.5.1"

[tool.black]
line-length = 119

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]
[tool.coverage.report]
fail_under = 100

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
